requirements.txt
pytest~=7.4.3
allure-pytest
selene>=2.0.0b4
selenium~=4.16.0